edit vaccine history, test event stats

+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/entities/VaccineProgramEntity.java
@@ -27,7 +27,7 @@ public class VaccineProgramEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "vaccine_id")
-    private Integer vaccineId;
+    private Long vaccineId;
 
     @Column(name = "vaccine_name")
     private String vaccineName;
